requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.4
django-filter==25.1
djangorestframework==3.16.0
et_xmlfile==2.0.0
gunicorn==23.0.0
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
